+++ R/autospc.R
@@ -23,7 +23,7 @@
 #' See \code{vignette("data-requirements", package = "autospc")} for more
 #' details.
 #' @param chart_type The type of chart you wish to plot. Must must have length one.
-#' Available options are: "XMR", "MR", "C", "C'", "P", "P'".
+#' Available options are: "XMR", "X", "MR", "C", "C'", "P", "P'".
 #' 
 #' ## Algorithm Parameters
 #' Parameters that control behaviour of the algorithm used to re-establish